requirements.txt
Django==5.2.5
djangorestframework==3.15.2
django-cors-headers==4.5.0
drf-spectacular==0.27.2
django-filter==24.3
